Integrates all user feedback with issue tracker (Fixes issue 3). Replaced footer requesting comments/question be emailed to Sal with links for:
Report a problem
Make a suggestion
Request a video
Make a comment
Ask a question

I was originally planning to use the Issue Tracker API to implement this
but it looked like that would be more complicated and would  
require the user to give our app permission to create and vote on issues 
for them across all Google Code projects.  Even though we would restrict 
ourselves to khanacademy, it would probably still be problematic for some
users.  It would also require the user to go through an extra page to give
us that permission which would be confusing.

My implementation uses a combination of issue tracker templates and query
urls on pages with (hopefully) clear instructions.

Testing of an earlier version by an 8yo was mostly successful and lead to
clearer instructions.



git-svn-id: http://khanacademy.googlecode.com/svn/trunk@24 5cfed7ce-a931-7a28-2923-4f78740e45b1