Screened empty text() calls, which removes the "undefined" hints in division and multiplication.
Adjusted graph sizes in a bunch of exercises to get rid of scrollbars, and also changed them to multiple_choice_template.

limits_2 won't try to graph a hint if it doesn't need to, because Raphael crashed on that, unlike ASCIIsvg. 

Dotted lines work, but I'm forcing them to use one style in the mean time; I'll fix it later if someone really needs different dotted line styles.

With this commit, all exercises are tested and working in IE to my knowledge except domain_of_a_function. The R (all real numbers) symbol still doesn't show up in that exercise in IE, but it does appear in other browsers.

The borders still appear around the graph area as well.



git-svn-id: http://khanacademy.googlecode.com/svn/trunk@1152 5cfed7ce-a931-7a28-2923-4f78740e45b1