Replace scratchpad library svg-edit with jsvectoreditor for IE support: https://code.google.com/p/jsvectoreditor/
Note that we use the "painter.html" and "painter/" files in the trunk, not the older files still in there. I've also cut out a lot of unnecessary tools and features just like Dean did with svg-edit, which we can add back in easily if desired. Also, I changed the default fill to "none" and stroke to 2px black.

Should already work on all browsers, and we add IE8 and possibly (not tested) iPad support immediately. IE7 support is coming next. This editor's a little bit less usable and worse-looking in some ways, particularly text entry, but it works.



git-svn-id: http://khanacademy.googlecode.com/svn/trunk@1260 5cfed7ce-a931-7a28-2923-4f78740e45b1